Make it harder to forget itemCount for FixedHeightScrollable
Previously, if a subclass of FixedHeightScrollable didn't set itemCount,
everything would silently work except that you wouldn't actully be able to
scroll. This CL makes it harder to forget to provide the itemCount by making it
an abstract getter. Now the analyzer will flag it as missing in subclasses and
we'll throw an exception at runtime if you forget it.

R=ianh@google.com, tonyg@google.com

Review URL: https://codereview.chromium.org/1219933004 .