Allow multiple callbacks in the VSyncProvider
It's possible for multiple calls to AwaitVSync to be pending while reloading
an application.  The Animator may get an AwaitVSync, a Reset, and then another
AwaitVSync before the first vsync completes.

If this happens, invoke both callbacks.  If we close the pipe, then the
Animator will no longer make progress.