Improving support for AOT builds of Fuchsia packages
The integration test works in debug/jit mode only, and building in
`profile` and `release` modes was not working, breaking some builders.

The workaround was to disable the test.

This PR will enable release/aot mode so it will build, and unblock both
the debug-mode version of the test (tested by CI) and release mode
builds of the repo.

Due to the request to unblock, this PR does not yet build packages in
`profile` mode that will run in Fuchsia, but it should at least build
without errors. Ongoing work is in progress to also support running in
release mode.