[Impeller] Generate a Metal library symbols file for shader debugging. (#39532)
Towards fixing https://github.com/flutter/flutter/issues/120326. This allows you
to specify the sources for the Impeller Metal Library. The user is meant to
navigate to `out/<engine_variant>/shaders/entity.metallib` when Xcode prompts
for sources.

Shader debugging with line by line profiling is still not a turnkey operation
though. This works fine on macOS. But on iOS, I had to bump up the min iOS
version level to iOS 12 at least and specify `-g` to the `metal` compiler for
the sources to show up with profiling. The performance of the shaders was
identical after the changes though so I suspect engine developers will patch in
these flags locally to develop and debug shaders.