[Impeller] Disable blending for Source blend mode (#43162)
All of the backends already respect this property for the first color
attachment in the pipeline. ~~Our Vulkan backend disables blending for a
whole pipeline if the first color attachment has blending turned off.~~
This has been under our nose but I haven't checked it until now.

HSR might not be working on some Vulkan & OpenGLES drivers because of
this.