[Impeller] Remove use of FML_DISALLOW_<FOO> macros in Impeller. (#47307)
Really not convinced this is easier to read but maybe I just need to get used to it. Besides, good to just stick to the style guide uniformly.

Followup from https://github.com/flutter/engine/pull/47278#discussion_r1371971876

Can do this for the rest of the engine and remove the macros if this is fine.