Restore use of the API 34 device in the Firebase Test Lab script and handle FTL infrastructure errors in the script (#50735)
This test had been temporarily moved to an API 33 device as a workaround for an FTL infrastructure issue (see https://github.com/flutter/engine/pull/50721)

This PR resumes use of a Pixel 8/API 34 device for the FTL tests.  It also retries the test if FTL returns an error code that is known to represent an FTL infrastructure error.  If the retries fail, then the test script will not block the engine tree if all failures were caused by infrastructure.

This is similar to what the CI recipes are doing in https://flutter.googlesource.com/recipes/+/a181878fde742dacd94afca04c5a4db1b3c30b91