[Impeller] fix incorrect padding/translation in drawVertices with texture coordinates. (#53746)
Fixes https://github.com/flutter/flutter/issues/151355

The coverage pad introduced by renderToSnapshot is being picked up by drawVertices with texture coordinates. The pad is showing up in the final results, which can appear to be a gap between elements.

Additionally: if the coverage computed from the texture coordinates does not include the origin, then make sure the coverage used for the snapshot is translated onscreen.